LayoutTests/platform/glib/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/onpaint-fires-post-intersection-observation-steps-expected.txt


FAIL intersection observations should be calculated before onpaint, but should be delivered after onpaint assert_array_equals: onpaint should fire before intersection observations are delivered. expected property 0 to be "onpaint" but got "intersection" (expected array ["onpaint", "intersection"] got ["intersection", "onpaint"])
